Rename post-summary → post-details-header
