+++ apps/web/src/components/domains/post/post-details-header/index.ts
@@ -1,0 +1,1 @@
+export * from './post-details-header';
