--- apps/web/src/components/domains/post/post-summary/index.ts
@@ -1,1 +1,0 @@
-export * from './post-summary';
